+++ CHANGELOG.md
@@ -1,3 +1,7 @@
-## [0.0.1] - 2019-06-15.
+## [0.0.2] - 2021-03-19.
 
 * Flutter plugin to create lightbox like in web apps.
+* Added support for Files and base64 encoded string to be read as bytes
+* Exported library files
+* Added more documentation
+* ImageType to specify the type of images passed to the widget